Move exp and logarithm modules into arithmetic
