+++ src/integer/z.rs
@@ -15,12 +15,10 @@ mod arithmetic;
 mod cmp;
 mod default;
 mod distance;
-mod exp;
 pub(crate) mod fmpz_helpers;
 mod from;
 mod gcd;
 mod lcm;
-mod logarithm;
 mod ownership;
 mod properties;
 mod sample;
