+++ src/integer/z/arithmetic.rs
@@ -11,6 +11,8 @@
 
 mod add;
 mod div;
+mod exp;
+mod logarithm;
 mod mul;
 mod pow;
 mod root;
