+++ src/rational/poly_over_q.rs
@@ -16,7 +16,6 @@ mod arithmetic;
 mod cmp;
 mod default;
 mod evaluate;
-mod exp;
 mod from;
 mod get;
 mod ownership;
