+++ src/rational/poly_over_q/arithmetic/exp.rs
@@ -8,7 +8,7 @@
 
 //! Implementations to approximate the exponential function using a [`PolyOverQ`] polynomial.
 
-use super::PolyOverQ;
+use crate::rational::PolyOverQ;
 use flint_sys::fmpq_poly::fmpq_poly_exp_series;
 use std::str::FromStr;
