+++ src/rational/q/arithmetic.rs
@@ -11,6 +11,7 @@
 
 mod add;
 mod div;
+mod exp;
 mod mul;
 mod pow;
 mod sub;
