+++ src/rational/q/arithmetic/exp.rs
@@ -6,15 +6,14 @@
 // the terms of the Mozilla Public License Version 2.0 as published by the
 // Mozilla Foundation. See <https://mozilla.org/en-US/MPL/2.0/>.
 
-//! Implementations to call the exponential function on a [`Z`] integer.
+//! Implementations to call the exponential function on a [`Q`].
 
-use super::Z;
 use crate::{
     rational::{PolyOverQ, Q},
     traits::Evaluate,
 };
 
-impl Z {
+impl Q {
     /// Computes e^self using taylor series approximation of the exponential function.
     ///
     /// Parameters:
@@ -25,12 +24,13 @@ impl Z {
     ///
     /// # Examples
     /// ```
-    /// use qfall_math::integer::Z;
+    /// use qfall_math::rational::Q;
+    /// use std::str::FromStr;
     ///
-    /// // sum_{k=0}^999 17^k/k!
-    /// let evaluation = Z::from(17).exp_taylor(1000_u32);
+    /// // sum_{k=0}^999 (17/3)^k/k!
+    /// let evaluation = Q::from_str("17/3").unwrap().exp_taylor(1000_u32);
     /// ```
-    pub fn exp_taylor(&self, length_taylor_polynomial: impl Into<u32>) -> Q {
+    pub fn exp_taylor(&self, length_taylor_polynomial: impl Into<u32>) -> Self {
         let exp_taylor_series = PolyOverQ::exp_function_taylor(length_taylor_polynomial);
         exp_taylor_series.evaluate(self)
     }
@@ -38,15 +38,15 @@ impl Z {
 
 #[cfg(test)]
 mod test_exp {
-    use crate::{integer::Z, rational::Q};
+    use crate::rational::Q;
     use std::str::FromStr;
 
     /// ensure that `0` is returned if the length `0` is provided
     #[test]
     fn zero_length() {
-        let z = Z::from(17);
+        let q = Q::from_str("17/3").unwrap();
 
-        assert_eq!(Q::default(), z.exp_taylor(0_u32));
+        assert_eq!(Q::default(), q.exp_taylor(0_u32));
     }
 
     /// test correct evaluation for some explicit values
@@ -54,15 +54,15 @@ mod test_exp {
     fn ten_length_value() {
         assert_eq!(
             Q::from_str("98641/36288").unwrap(),
-            Z::ONE.exp_taylor(10_u32)
+            Q::from_str("1").unwrap().exp_taylor(10_u32)
         );
         assert_eq!(
-            Q::from_str("22471/1120").unwrap(),
-            Z::from(3).exp_taylor(10_u32)
+            Q::from_str("2492063827/1785641760").unwrap(),
+            Q::from_str("1/3").unwrap().exp_taylor(10_u32)
         );
         assert_eq!(
-            Q::from_str("83/2240").unwrap(),
-            Z::from(-3).exp_taylor(10_u32)
+            Q::from_str("5729869/11160261").unwrap(),
+            Q::from_str("-2/3").unwrap().exp_taylor(10_u32)
         );
     }
 }
